backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==1.4.53
asyncpg>=0.29.0
databases==0.8.0
pydantic==2.9.2
python-dotenv==1.0.1
cryptography==43.0.3
httpx==0.27.2

